+++ src/main/cxx/core/equations.cxx
@@ -1,5 +1,5 @@
-/// @file exitcodes.h
-/// Application error exit codes.
+/// @file util.cxx
+/// Utilities
 /// Copyright (c) 2017 Ralf Quast
 ///
 /// Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -19,50 +19,30 @@
 /// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 /// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 /// SOFTWARE.
-#ifndef ESPECIA_EXITCODES_H
-#define ESPECIA_EXITCODES_H
 
-namespace especia {
+#include "equations.h"
 
-    /**
-     * Application error exit codes.
-     */
-    class Exit_Codes {
-    public:
-        /**
-         * The optimization stopped due to an underflow of the mutation variance.
-         */
-        static const int OPTIMIZATION_UNDERFLOW = 1;
+void especia::Equations::birch94(const R_type &x, R_type &y, R_type &z) {
+    const R_type n = 1.0 + 8.34254E-05 + 2.406147E-08 / (130.0E-06 - x * x) + 1.5998E-10 / (38.9E-06 - x * x);
+    const R_type m = (4.812294E-08 * x) / sqr(130.0E-06 - x * x) + (3.1996E-10 * x) / sqr(38.9E-06 - x * x);
 
-        /**
-         * The optimization stopped and failed to reach the required accuracy goal.
-         */
-        static const int OPTIMIZATION_STOPPED = 2;
-
-        /**
-         * A logical error occurred.
-         */
-        static const int LOGICAL_ERROR = 10;
-
-        /**
-         * A runtime error occurred.
-         */
-        static const int RUNTIME_ERROR = 11;
-
-        /**
-         * An unspecific exception occurred.
-         */
-        static const int UNSPECIFIC_EXCEPTION = 12;
-
-    private:
-        /**
-         * Private constructor prevents instantiation.
-         */
-        Exit_Codes() {
-
-        }
-    };
-
+    y = x * n;
+    z = n + x * m;
 }
 
-#endif // ESPECIA_EXITCODES_H
+void especia::Equations::edlen53(const R_type &x, R_type &y, R_type &z) {
+    const R_type n = 1.0 + 6.43280E-05 + 2.5540E-10 / (0.0000410 - x * x) + 2.949810E-08 / (0.000146 - x * x);
+    const R_type m = (5.1080E-10 * x) / sqr(0.0000410 - x * x) + (5.89962E-08 * x) / sqr(0.000146 - x * x);
+
+    y = x * n;
+    z = n + x * m;
+}
+
+void especia::Equations::edlen66(const R_type &x, R_type &y, R_type &z) {
+    const R_type n = 1.0 + 8.34213E-05 + 1.5997E-10 / (0.0000389 - x * x) + 2.406030E-08 / (0.000130 - x * x);
+    const R_type m = (3.1994E-10 * x) / sqr(0.0000389 - x * x) + (4.81206E-08 * x) / sqr(0.000130 - x * x);
+
+    y = x * n;
+    z = n + x * m;
+}
+