+++ src/main/cxx/util/helicorr.cxx
@@ -19,7 +19,8 @@
 /// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 /// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 /// SOFTWARE.
-#include <cmath>
+#include <exception>
+#include <stdexcept>
 
 #include "../core/base.h"
 #include "../core/dataio.h"