+++ NEWS.md
@@ -4,9 +4,9 @@ The main programs were renamed and a few utility programs were added.
 The source code, documentation, and ancillary files were revised.
 Version 2017.1 provides new features and enhancements:
 
-* Equations to convert photon wavelength in vacuum to photon wavelength
-in standard air (and vice versa) were added to the core API (Edlén 1953,
-1966; Birch & Downs 1994).
+* Equations and equation solvers to convert photon wavelength in vacuum
+to photon wavelength in standard air (and vice versa) were added to the
+core API (Edlén 1953, 1966; Birch & Downs 1994).
 * Especia uses standard C++-11 multithreading, if your compiler does
 not support Open Multiprocessing directives.
 * The spectral resolution of the instrument is expressed in units of